DashBoard/requirements.txt
streamlit==0.79.0
pandas==1.1.5
numpy==1.19.4
matplotlib==3.3.3
nltk==3.5
Pillow==8.2.0
plotly==4.14.3
wordcloud==1.8.1
